Refactor model API methods to a separate file
